refactor: extract out checkout creation handler
