+++ firestore-stripe-payments/functions/__tests__/jest.global-setup.ts
@@ -11,6 +11,12 @@ declare global {
 }
 
 export default async function globalSetup() {
+  // Skip Stripe setup for non-integration tests
+  if (process.env.INTEGRATION_TEST !== 'true') {
+    console.log('Skipping Stripe setup for unit tests');
+    return;
+  }
+
   // Load all environment variables
   setupEnvironment();
