+++ firestore-stripe-payments/functions/package.json
@@ -11,7 +11,8 @@
     "build:watch": "tsc -p tsconfig.build.json --watch",
     "compile": "tsc",
     "generate-readme": "firebase ext:info .. --markdown > ../README.md",
-    "test": "cd ../_emulator && firebase emulators:exec -P demo-project \"jest --config=../functions/jest.config.js\"",
+    "test": "cd ../_emulator && firebase emulators:exec -P demo-project \"INTEGRATION_TEST=true jest --config=../functions/jest.config.js\"",
+    "test:unit": "jest --config=jest.config.js",
     "start:emulator": "cd ../_emulator && firebase emulators:start -P demo-project",
     "exec:emulator": "cd ../_emulator && firebase emulators:exec \"../runTests.sh\" -P demo-project",
     "exec:emulator:watch": "cd ../_emulator && firebase emulators:exec \"../runTestsWatch.sh\" -P demo-project",
@@ -19,7 +20,7 @@
   },
   "author": "Stripe (https://stripe.com/)",
   "license": "Apache-2.0",
- "dependencies": {
+  "dependencies": {
     "@types/node": "^18.15.3",
     "firebase-admin": "^11.10.1",
     "firebase-functions": "^3.21.2",
@@ -29,13 +30,13 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^6.0.0",
-    "@types/chai": "^4.1.6",
     "@types/express": "^4.17.13",
     "@types/jest": "^29.5.14",
     "concurrently": "^7.0.0",
     "dotenv": "^16.0.0",
     "envfile": "^6.17.0",
     "firebase-functions-test": "^0.3.3",
+
     "jest": "^29.7.0",
     "localtunnel": "^2.0.2",
     "mocked-env": "^1.3.5",
