+++ firestore-stripe-payments/functions/src/config.ts
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import Stripe from 'stripe';
+
 export interface ExtensionConfig {
   stripeSecretKey: string;
   stripeWebhookSecret?: string;
@@ -36,5 +38,17 @@ const config: ExtensionConfig = {
   minCheckoutInstances:
     Number(process.env.CREATE_CHECKOUT_SESSION_MIN_INSTANCES) ?? 0,
 };
+
+export const apiVersion = '2022-11-15';
+
+export const stripe = new Stripe(config.stripeSecretKey, {
+  apiVersion,
+  // Register extension as a Stripe plugin
+  // https://stripe.com/docs/building-plugins#setappinfo
+  appInfo: {
+    name: 'Firebase Invertase firestore-stripe-payments',
+    version: '0.3.5',
+  },
+});
 
 export default config;
